requirements.txt
# Core Django
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-jazzmin==2.6.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7
djangorestframework-simplejwt==5.3.1
argon2-cffi==23.1.0

# Celery & Task Queue
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
kombu==5.3.4

# API Documentation
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1

# Monitoring & Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-django==0.43b0
opentelemetry-instrumentation-psycopg2==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-exporter-otlp==1.22.0
sentry-sdk==1.39.2

# Utilities
python-dotenv==1.0.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.2.0
pytz==2023.3.post1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0
freezegun==1.4.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
django-stubs==4.2.7

# Payment Integration
stripe==7.9.0

# Validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.27
